Update image rendering

+++ src/lib/Helper/Image/AbstractImageHelper.php
@@ -43,12 +43,12 @@ abstract class AbstractImageHelper {
         int $maxWidth,
         int $maxHeight
     ) {
-        $heightWidthRatio = $height / $width;
-        $widthHeightRatio = $width / $height;
+        $heightWidthRatio = (int) $height / $width;
+        $widthHeightRatio = (int) $width / $height;
 
         $size = [
-            'width'      => $minWidth,
-            'height'     => $minWidth * $heightWidthRatio,
+            'width'      => (int) $minWidth,
+            'height'     => (int) $minWidth * $heightWidthRatio,
             'cropX'      => 0,
             'cropY'      => 0,
             'cropWidth'  => 0,
@@ -57,36 +57,36 @@ abstract class AbstractImageHelper {
         ];
 
         if($minHeight !== 0 && $size['height'] < $minHeight) {
-            $size['width']  = $minHeight * $widthHeightRatio;
-            $size['height'] = $minHeight;
+            $size['width']  = (int) $minHeight * $widthHeightRatio;
+            $size['height'] = (int) $minHeight;
 
             if($maxWidth !== 0 && $size['width'] > $maxWidth) {
-                $size['cropX']      = ($size['width'] - $maxWidth) / 2;
-                $size['cropWidth']  = $maxWidth;
-                $size['cropHeight'] = $size['height'];
+                $size['cropX']      = (int) ($size['width'] - $maxWidth) / 2;
+                $size['cropWidth']  = (int) $maxWidth;
+                $size['cropHeight'] = (int) $size['height'];
                 $size['cropNeeded'] = true;
             }
         } else if($maxHeight !== 0 && $size['height'] > $maxHeight) {
-            $size['width'] = $minHeight * $widthHeightRatio;
+            $size['width'] = (int) $minHeight * $widthHeightRatio;
 
             if($maxWidth !== 0 && $size['width'] > $maxWidth) {
-                $size['width']      = $maxWidth;
-                $size['height']     = $maxWidth * $heightWidthRatio;
-                $size['cropWidth']  = $size['width'];
-                $size['cropHeight'] = $maxHeight;
+                $size['width']      = (int) $maxWidth;
+                $size['height']     = (int) $maxWidth * $heightWidthRatio;
+                $size['cropWidth']  = (int) $size['width'];
+                $size['cropHeight'] = (int) $maxHeight;
                 $size['cropNeeded'] = true;
             } else if($size['width'] < $minWidth) {
-                $size['width']      = $minWidth;
-                $size['height']     = $minWidth * $heightWidthRatio;
-                $size['cropWidth']  = $size['width'];
-                $size['cropHeight'] = $maxHeight;
+                $size['width']      = (int) $minWidth;
+                $size['height']     = (int) $minWidth * $heightWidthRatio;
+                $size['cropWidth']  = (int) $size['width'];
+                $size['cropHeight'] = (int) $maxHeight;
                 $size['cropNeeded'] = true;
             }
         }
 
         if($size['width'] === 0 && $size['height'] === 0) {
-            $size['width']  = $width;
-            $size['height'] = $height;
+            $size['width']  = (int) $width;
+            $size['height'] = (int) $height;
         }
 
         return $size;
